pyproject.toml
[project]
name = "tutorring-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.1.1",
    "fastapi>=0.119.0",
    "gtts>=2.5.4",
    "openai-agents>=0.3.3",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "speechrecognition>=3.14.3",
    "websockets>=15.0.1",
]
